Split Compiler Tests
