+++ engine/baml-compiler/src/codegen.rs
@@ -1596,2129 +1596,3 @@ impl<'g> HirCompiler<'g> {
         loop_info.break_patch_list
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use baml_vm::EvalStack;
-
-    use super::*;
-    use crate::test::ast;
-
-    /// Helper struct for testing bytecode compilation.
-    struct Program {
-        source: &'static str,
-        expected: Vec<(&'static str, Vec<Instruction>)>,
-    }
-
-    /// Helper function to assert that source code compiles to expected bytecode
-    /// instructions.
-    fn assert_compiles(input: Program) -> anyhow::Result<()> {
-        let ast = ast(input.source)?;
-
-        let BamlVmProgram {
-            objects, globals, ..
-        } = compile(&ast)?;
-
-        // Create a map of function name to function for easy lookup
-        let functions: std::collections::HashMap<&str, &baml_vm::Function> = objects
-            .iter()
-            .filter_map(|obj| match obj {
-                Object::Function(f) => Some((f.name.as_str(), f)),
-                _ => None,
-            })
-            .collect();
-
-        // Check each expected function
-        for (function_name, expected_instructions) in input.expected {
-            let function = functions
-                .get(function_name)
-                .ok_or_else(|| anyhow::anyhow!("function '{}' not found", function_name))?;
-
-            eprintln!(
-                "---- fn {function_name}() ----\n{}",
-                baml_vm::debug::display_bytecode(
-                    function,
-                    &EvalStack::default(),
-                    &objects,
-                    &globals,
-                    true
-                )
-            );
-
-            assert_eq!(
-                function.bytecode.instructions, expected_instructions,
-                "Bytecode mismatch for function '{function_name}'"
-            );
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn return_function_call() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: "
-                fn one() -> int {
-                    1
-                }
-
-                fn main() -> int {
-                    one()
-                }
-            ",
-            expected: vec![
-                ("one", vec![Instruction::LoadConst(0), Instruction::Return]),
-                (
-                    "main",
-                    vec![
-                        Instruction::LoadGlobal(GlobalIndex::from_raw(0)),
-                        Instruction::Call(0),
-                        Instruction::Return,
-                    ],
-                ),
-            ],
-        })
-    }
-
-    #[test]
-    fn call_function() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: "
-                fn two() -> int {
-                    2
-                }
-
-                fn main() -> int {
-                    let a = two();
-                    a
-                }
-            ",
-            expected: vec![
-                ("two", vec![Instruction::LoadConst(0), Instruction::Return]),
-                (
-                    "main",
-                    vec![
-                        Instruction::LoadGlobal(GlobalIndex::from_raw(0)),
-                        Instruction::Call(0),
-                        Instruction::LoadVar(1),
-                        Instruction::Return,
-                    ],
-                ),
-            ],
-        })
-    }
-
-    #[test]
-    fn if_else_return_expr() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: "
-                fn main(b: bool) -> int {
-                    if (b) { 1 } else { 2 }
-                }
-            ",
-            expected: vec![(
-                "main",
-                vec![
-                    Instruction::LoadVar(1),
-                    Instruction::JumpIfFalse(4),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(0),
-                    Instruction::Jump(3),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(1),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn if_else_return_expr_with_locals() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: "
-                fn main(b: bool) -> int {
-                    if (b) {
-                        let a = 1;
-                        a
-                    } else {
-                        let a = 2;
-                        a
-                    }
-                }
-            ",
-            expected: vec![(
-                "main",
-                vec![
-                    Instruction::LoadVar(1),
-                    Instruction::JumpIfFalse(6),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(0),
-                    Instruction::LoadVar(2),
-                    Instruction::PopReplace(1),
-                    Instruction::Jump(5),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(1),
-                    Instruction::LoadVar(2),
-                    Instruction::PopReplace(1),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn if_else_assignment() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: "
-                fn main(b: bool) -> int {
-                    let i = if (b) { 1 } else { 2 };
-                    i
-                }
-            ",
-            expected: vec![(
-                "main",
-                vec![
-                    Instruction::LoadVar(1),
-                    Instruction::JumpIfFalse(4),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(0),
-                    Instruction::Jump(3),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(1),
-                    Instruction::LoadVar(2),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn if_else_assignment_with_locals() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: "
-                fn main(b: bool) -> int {
-                    let i = if (b) {
-                        let a = 1;
-                        a
-                    } else {
-                        let a = 2;
-                        a
-                    };
-
-                    i
-                }
-            ",
-            expected: vec![(
-                "main",
-                vec![
-                    Instruction::LoadVar(1),
-                    Instruction::JumpIfFalse(6),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(0),
-                    Instruction::LoadVar(2),
-                    Instruction::PopReplace(1),
-                    Instruction::Jump(5),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(1),
-                    Instruction::LoadVar(2),
-                    Instruction::PopReplace(1),
-                    Instruction::LoadVar(2),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn if_else_normal_statement() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: "
-                fn identity(i: int) -> int {
-                    i
-                }
-
-                fn main(b: bool) -> int {
-                    let a = 1;
-
-                    if (b) {
-                        let x = 1;
-                        let y = 2;
-                        identity(x);
-                    } else {
-                        let x = 3;
-                        let y = 4;
-                        identity(y);
-                    }
-
-                    a
-                }
-            ",
-            expected: vec![(
-                "main",
-                vec![
-                    Instruction::LoadConst(0),
-                    Instruction::LoadVar(1),
-                    Instruction::JumpIfFalse(10),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(1),
-                    Instruction::LoadConst(2),
-                    Instruction::LoadGlobal(GlobalIndex::from_raw(0)),
-                    Instruction::LoadVar(3),
-                    Instruction::Call(1),
-                    Instruction::Pop(1),
-                    Instruction::Pop(2),
-                    Instruction::Jump(9),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(3),
-                    Instruction::LoadConst(4),
-                    Instruction::LoadGlobal(GlobalIndex::from_raw(0)),
-                    Instruction::LoadVar(4),
-                    Instruction::Call(1),
-                    Instruction::Pop(1),
-                    Instruction::Pop(2),
-                    Instruction::LoadVar(2),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn else_if_return_expr() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: "
-                fn main(a: bool, b: bool) -> int {
-                    if (a) {
-                        1
-                    } else if (b) {
-                        2
-                    } else {
-                        3
-                    }
-                }
-            ",
-            expected: vec![(
-                "main",
-                vec![
-                    Instruction::LoadVar(1),
-                    Instruction::JumpIfFalse(4),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(0),
-                    Instruction::Jump(9),
-                    Instruction::Pop(1),
-                    Instruction::LoadVar(2),
-                    Instruction::JumpIfFalse(4),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(1),
-                    Instruction::Jump(3),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(2),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn else_if_return_expr_with_locals() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: "
-                fn main(a: bool, b: bool) -> int {
-                    if (a) {
-                        let x = 1;
-                        x
-                    } else if (b) {
-                        let y = 2;
-                        y
-                    } else {
-                        let z = 3;
-                        z
-                    }
-                }
-            ",
-            expected: vec![(
-                "main",
-                vec![
-                    Instruction::LoadVar(1),
-                    Instruction::JumpIfFalse(6),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(0),
-                    Instruction::LoadVar(3),
-                    Instruction::PopReplace(1),
-                    Instruction::Jump(13),
-                    Instruction::Pop(1),
-                    Instruction::LoadVar(2),
-                    Instruction::JumpIfFalse(6),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(1),
-                    Instruction::LoadVar(3),
-                    Instruction::PopReplace(1),
-                    Instruction::Jump(5),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(2),
-                    Instruction::LoadVar(3),
-                    Instruction::PopReplace(1),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn else_if_assignment() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: "
-                fn main(a: bool, b: bool) -> int {
-                    let result = if (a) {
-                        1
-                    } else if (b) {
-                        2
-                    } else {
-                        3
-                    };
-
-                    result
-                }
-            ",
-            expected: vec![(
-                "main",
-                vec![
-                    Instruction::LoadVar(1),
-                    Instruction::JumpIfFalse(4),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(0),
-                    Instruction::Jump(9),
-                    Instruction::Pop(1),
-                    Instruction::LoadVar(2),
-                    Instruction::JumpIfFalse(4),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(1),
-                    Instruction::Jump(3),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(2),
-                    Instruction::LoadVar(3),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn else_if_assignment_with_locals() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: "
-                fn main(a: bool, b: bool) -> int {
-                    let result = if (a) {
-                        let x = 1;
-                        x
-                    } else if (b) {
-                        let y = 2;
-                        y
-                    } else {
-                        let z = 3;
-                        z
-                    };
-
-                    result
-                }
-            ",
-            expected: vec![(
-                "main",
-                vec![
-                    Instruction::LoadVar(1),
-                    Instruction::JumpIfFalse(6),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(0),
-                    Instruction::LoadVar(3),
-                    Instruction::PopReplace(1),
-                    Instruction::Jump(13),
-                    Instruction::Pop(1),
-                    Instruction::LoadVar(2),
-                    Instruction::JumpIfFalse(6),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(1),
-                    Instruction::LoadVar(3),
-                    Instruction::PopReplace(1),
-                    Instruction::Jump(5),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(2),
-                    Instruction::LoadVar(3),
-                    Instruction::PopReplace(1),
-                    Instruction::LoadVar(3),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn array_constructor() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: "
-                fn main() -> int[] {
-                    let a = [1, 2, 3];
-                    a
-                }
-            ",
-            expected: vec![(
-                "main",
-                vec![
-                    Instruction::LoadConst(0),
-                    Instruction::LoadConst(1),
-                    Instruction::LoadConst(2),
-                    Instruction::AllocArray(3),
-                    Instruction::LoadVar(1),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn class_constructor() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: "
-                class Point {
-                    x int
-                    y int
-                }
-
-                fn main() -> Point {
-                    let p = Point { x: 1, y: 2 };
-                    p
-                }
-            ",
-            expected: vec![(
-                "main",
-                vec![
-                    Instruction::AllocInstance(ObjectIndex::from_raw(2)),
-                    Instruction::Copy(0),
-                    Instruction::LoadConst(0),
-                    Instruction::StoreField(0),
-                    Instruction::Copy(0),
-                    Instruction::LoadConst(1),
-                    Instruction::StoreField(1),
-                    Instruction::LoadVar(1),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn class_constructor_with_spread_operator() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: r#"
-                class Point {
-                    x int
-                    y int
-                    z int
-                    w int
-                }
-
-                fn default_point() -> Point {
-                    Point { x: 0, y: 0, z: 0, w: 0 }
-                }
-
-                fn main() -> Point {
-                    let p = Point { x: 1, y: 2, ..default_point() };
-                    p
-                }
-            "#,
-            expected: vec![(
-                "main",
-                vec![
-                    Instruction::AllocInstance(ObjectIndex::from_raw(3)),
-                    Instruction::Copy(0),
-                    Instruction::LoadConst(0),
-                    Instruction::StoreField(0),
-                    Instruction::Copy(0),
-                    Instruction::LoadConst(1),
-                    Instruction::StoreField(1),
-                    Instruction::LoadGlobal(GlobalIndex::from_raw(0)),
-                    Instruction::Call(0),
-                    Instruction::Copy(1), // Copy instance from under spread
-                    Instruction::Copy(1), // Copy spread from under instance
-                    Instruction::LoadField(2),
-                    Instruction::StoreField(2),
-                    Instruction::Copy(1), // Copy instance from under spread
-                    Instruction::Copy(1), // Copy spread from under instance
-                    Instruction::LoadField(3),
-                    Instruction::StoreField(3),
-                    Instruction::Pop(1),
-                    Instruction::LoadVar(1),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn class_constructor_with_spread_operator_does_not_break_locals() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: r#"
-                class Point {
-                    x int
-                    y int
-                    z int
-                    w int
-                }
-
-                fn default_point() -> Point {
-                    Point { x: 0, y: 0, z: 0, w: 0 }
-                }
-
-                fn main() -> int {
-                    let p = Point { x: 1, y: 2, ..default_point() };
-                    let x = 0;
-                    x
-                }
-            "#,
-            expected: vec![(
-                "main",
-                vec![
-                    Instruction::AllocInstance(ObjectIndex::from_raw(3)),
-                    Instruction::Copy(0),
-                    Instruction::LoadConst(0),
-                    Instruction::StoreField(0),
-                    Instruction::Copy(0),
-                    Instruction::LoadConst(1),
-                    Instruction::StoreField(1),
-                    Instruction::LoadGlobal(GlobalIndex::from_raw(0)),
-                    Instruction::Call(0),
-                    Instruction::Copy(1), // Copy instance from under spread
-                    Instruction::Copy(1), // Copy spread from under instance
-                    Instruction::LoadField(2),
-                    Instruction::StoreField(2),
-                    Instruction::Copy(1), // Copy instance from under spread
-                    Instruction::Copy(1), // Copy spread from under instance
-                    Instruction::LoadField(3),
-                    Instruction::StoreField(3),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(2),
-                    Instruction::LoadVar(2),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn function_returning_string() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: r#"
-                fn main() -> string {
-                    "hello"
-                }
-            "#,
-            expected: vec![("main", vec![Instruction::LoadConst(0), Instruction::Return])],
-        })
-    }
-
-    #[test]
-    fn block_expr() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: "
-                fn main() -> int {
-                    let a = {
-                        let b = 1;
-                        b
-                    };
-
-                    a
-                }
-            ",
-            expected: vec![(
-                "main",
-                vec![
-                    Instruction::LoadConst(0),
-                    Instruction::LoadVar(1),
-                    Instruction::PopReplace(1),
-                    Instruction::LoadVar(1),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn locals_in_scope() -> anyhow::Result<()> {
-        let ast = ast(r#"
-            fn main() -> int {
-                let x = 0;
-
-                let a = {
-                    let y = 0;
-
-                    let b = {
-                        let c = 1;
-                        let d = 2;
-                        [c, d]
-                    };
-                    let e = {
-                        let f = 4;
-                        let g = 5;
-                        [f, g]
-                    };
-
-                    [b, e]
-                };
-
-                let h = {
-                    let z = 0;
-
-                    let i = {
-                        let w = 0;
-                        let j = 8;
-                        [w, j]
-                    };
-
-                    [i]
-                };
-
-                0
-            }
-        "#)?;
-
-        let BamlVmProgram {
-            objects,
-            resolved_function_names,
-            globals,
-            ..
-        } = compile(&ast)?;
-
-        let main = objects[resolved_function_names["main"].0].as_function()?;
-        baml_vm::debug::disassemble(main, &EvalStack::default(), &objects, &globals);
-
-        let expected_locals_in_scope = [
-            vec!["<fn main>", "x", "a", "h"],
-            vec!["<fn main>", "x", "y", "b", "e"],
-            vec!["<fn main>", "x", "y", "c", "d"],
-            vec!["<fn main>", "x", "y", "b", "f", "g"],
-            vec!["<fn main>", "x", "a", "z", "i"],
-            vec!["<fn main>", "x", "a", "z", "w", "j"],
-        ];
-
-        assert_eq!(
-            main.locals_in_scope,
-            expected_locals_in_scope
-                .iter()
-                .map(|scope| scope.iter().map(ToString::to_string).collect::<Vec<_>>())
-                .collect::<Vec<_>>()
-        );
-
-        Ok(())
-    }
-
-    #[test]
-    fn mutable_variables() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: r#"
-                fn DeclareMutableInFunction(x: int) -> int {
-
-                    let y = 3;
-
-                    y = 5;
-
-                    y
-                }
-
-                fn MutableInArg(x: int) -> int {
-                    x = 3;
-                    x
-                }
-            "#,
-            expected: vec![
-                (
-                    "DeclareMutableInFunction",
-                    vec![
-                        Instruction::LoadConst(0),
-                        Instruction::LoadConst(1),
-                        Instruction::StoreVar(2),
-                        Instruction::LoadVar(2),
-                        Instruction::Return,
-                    ],
-                ),
-                (
-                    "MutableInArg",
-                    vec![
-                        Instruction::LoadConst(0),
-                        Instruction::StoreVar(1),
-                        Instruction::LoadVar(1),
-                        Instruction::Return,
-                    ],
-                ),
-            ],
-        })
-    }
-
-    #[test]
-    fn basic_and() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: r#"
-                fn ret_bool() -> bool {
-                    true
-                }
-
-                fn main() -> bool {
-                    true && ret_bool()
-                }
-            "#,
-            expected: vec![(
-                "main",
-                vec![
-                    Instruction::LoadConst(0),
-                    Instruction::JumpIfFalse(4),
-                    Instruction::Pop(1),
-                    Instruction::LoadGlobal(GlobalIndex::from_raw(0)),
-                    Instruction::Call(0),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn basic_or() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: r#"
-                fn ret_bool() -> bool {
-                    true
-                }
-
-                fn main() -> bool {
-                    true || ret_bool()
-                }
-            "#,
-            expected: vec![(
-                "main",
-                vec![
-                    Instruction::LoadConst(0),
-                    Instruction::JumpIfFalse(2),
-                    Instruction::Jump(4),
-                    Instruction::Pop(1),
-                    Instruction::LoadGlobal(GlobalIndex::from_raw(0)),
-                    Instruction::Call(0),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn basic_add() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: r#"
-                fn main() -> int {
-                    let a = 1 + 2;
-                    a
-                }
-            "#,
-            expected: vec![(
-                "main",
-                vec![
-                    Instruction::LoadConst(0),
-                    Instruction::LoadConst(1),
-                    Instruction::BinOp(BinOp::Add),
-                    Instruction::LoadVar(1),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn basic_assign_add() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: r#"
-                fn main() -> int {
-                    let x = 1;
-                    x += 2;
-                    x
-                }
-            "#,
-            expected: vec![(
-                "main",
-                vec![
-                    Instruction::LoadConst(0),
-                    Instruction::LoadVar(1),
-                    Instruction::LoadConst(1),
-                    Instruction::BinOp(BinOp::Add),
-                    Instruction::StoreVar(1),
-                    Instruction::LoadVar(1),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn while_loop_gcd() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: r#"
-                fn GCD(a: int, b: int) -> int {
-                    while (a != b) {
-                        if (a > b) {
-                            a = a - b;
-                        } else {
-                            b = b - a;
-                        }
-                    }
-
-                    a
-                }
-            "#,
-            expected: vec![(
-                "GCD",
-                vec![
-                    Instruction::LoadVar(1),
-                    Instruction::LoadVar(2),
-                    Instruction::CmpOp(CmpOp::NotEq),
-                    Instruction::JumpIfFalse(18),
-                    Instruction::Pop(1),
-                    Instruction::LoadVar(1),
-                    Instruction::LoadVar(2),
-                    Instruction::CmpOp(CmpOp::Gt),
-                    Instruction::JumpIfFalse(7),
-                    Instruction::Pop(1),
-                    Instruction::LoadVar(1),
-                    Instruction::LoadVar(2),
-                    Instruction::BinOp(BinOp::Sub),
-                    Instruction::StoreVar(1),
-                    Instruction::Jump(6),
-                    Instruction::Pop(1),
-                    Instruction::LoadVar(2),
-                    Instruction::LoadVar(1),
-                    Instruction::BinOp(BinOp::Sub),
-                    Instruction::StoreVar(2),
-                    Instruction::Jump(-20),
-                    Instruction::Pop(1),
-                    Instruction::LoadVar(1),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    // This tests that we don't emit POP_REPLACE for if expressions when they
-    // do not return values.
-    #[test]
-    fn nested_block_expr_with_ending_normal_if() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: "
-                fn main() -> int {
-                    let a = 1;
-
-                    {
-                        let b = 2;
-                        let c = 3;
-                        a = b + c;
-
-                        if (a == 5) {
-                            a = 10;
-                        }
-                    }
-
-                    a
-                }
-            ",
-            expected: vec![(
-                "main",
-                vec![
-                    Instruction::LoadConst(0),
-                    Instruction::LoadConst(1),
-                    Instruction::LoadConst(2),
-                    Instruction::LoadVar(2),
-                    Instruction::LoadVar(3),
-                    Instruction::BinOp(BinOp::Add),
-                    Instruction::StoreVar(1),
-                    Instruction::LoadVar(1),
-                    Instruction::LoadConst(3),
-                    Instruction::CmpOp(CmpOp::Eq),
-                    Instruction::JumpIfFalse(5),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(4),
-                    Instruction::StoreVar(1),
-                    Instruction::Jump(2),
-                    Instruction::Pop(1),
-                    Instruction::Pop(2),
-                    Instruction::LoadVar(1),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    // This tests that we don't emit POP_REPLACE for if expressions when they
-    // do not return values.
-    #[test]
-    fn while_loop_with_ending_if() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: "
-                fn main() -> int {
-                    let a = 1;
-
-                    while (a < 5) {
-                        a += 1;
-
-                        if (a == 2) {
-                            break;
-                        }
-                    }
-
-                    a
-                }
-            ",
-            expected: vec![(
-                "main",
-                vec![
-                    Instruction::LoadConst(0),
-                    Instruction::LoadVar(1),
-                    Instruction::LoadConst(1),
-                    Instruction::CmpOp(CmpOp::Lt),
-                    Instruction::JumpIfFalse(15),
-                    Instruction::Pop(1),
-                    Instruction::LoadVar(1),
-                    Instruction::LoadConst(2),
-                    Instruction::BinOp(BinOp::Add),
-                    Instruction::StoreVar(1),
-                    Instruction::LoadVar(1),
-                    Instruction::LoadConst(3),
-                    Instruction::CmpOp(CmpOp::Eq),
-                    Instruction::JumpIfFalse(4),
-                    Instruction::Pop(1),
-                    Instruction::Jump(5),
-                    Instruction::Jump(2),
-                    Instruction::Pop(1),
-                    Instruction::Jump(-17),
-                    Instruction::Pop(1),
-                    Instruction::LoadVar(1),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn break_factorial() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: r#"
-                fn Factorial(limit: int) -> int {
-                    let result = 1;
-
-                    while (true) {
-                        if (limit == 0) {
-                            break;
-                        }
-                        result = result * limit;
-                        limit = limit - 1;
-                    }
-
-                    result
-                }
-            "#,
-            expected: vec![(
-                "Factorial",
-                vec![
-                    // let result = 1;
-                    Instruction::LoadConst(0),
-                    // while true { ... }
-                    Instruction::LoadConst(1),
-                    Instruction::JumpIfFalse(19),
-                    Instruction::Pop(1),
-                    // if limit == 0 { break; }
-                    Instruction::LoadVar(1),
-                    Instruction::LoadConst(2),
-                    Instruction::CmpOp(CmpOp::Eq),
-                    Instruction::JumpIfFalse(4),
-                    Instruction::Pop(1),
-                    Instruction::Jump(13),
-                    Instruction::Jump(2),
-                    Instruction::Pop(1),
-                    // result = result * limit;
-                    Instruction::LoadVar(2),
-                    Instruction::LoadVar(1),
-                    Instruction::BinOp(BinOp::Mul),
-                    Instruction::StoreVar(2),
-                    // limit = limit - 1;
-                    Instruction::LoadVar(1),
-                    Instruction::LoadConst(3),
-                    Instruction::BinOp(BinOp::Sub),
-                    Instruction::StoreVar(1),
-                    // loop back and exit
-                    Instruction::Jump(-19),
-                    Instruction::Pop(1),
-                    // return result
-                    Instruction::LoadVar(2),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn continue_factorial() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: r#"
-                fn Factorial(limit: int) -> int {
-                    let result = 1;
-
-                    // used to make the loop break without relying on `break` implementation.
-                    let should_continue = true;
-                    while (should_continue) {
-                        result = result * limit;
-                        limit = limit - 1;
-
-                        if (limit != 0) {
-                            continue;
-                        } else {
-                            should_continue = false;
-                        }
-                    }
-
-                    result
-                }
-            "#,
-            expected: vec![(
-                "Factorial",
-                vec![
-                    // let result = 1;
-                    Instruction::LoadConst(0),
-                    // let should_continue = true;
-                    Instruction::LoadConst(1),
-                    // while should_continue { ... }
-                    Instruction::LoadVar(3),
-                    Instruction::JumpIfFalse(21),
-                    Instruction::Pop(1),
-                    // result = result * limit;
-                    Instruction::LoadVar(2),
-                    Instruction::LoadVar(1),
-                    Instruction::BinOp(BinOp::Mul),
-                    Instruction::StoreVar(2),
-                    // limit = limit - 1;
-                    Instruction::LoadVar(1),
-                    Instruction::LoadConst(2),
-                    Instruction::BinOp(BinOp::Sub),
-                    Instruction::StoreVar(1),
-                    // if limit != 0 { continue; } else { should_continue = false; }
-                    Instruction::LoadVar(1),
-                    Instruction::LoadConst(3),
-                    Instruction::CmpOp(CmpOp::NotEq),
-                    Instruction::JumpIfFalse(4),
-                    Instruction::Pop(1),
-                    Instruction::Jump(5),
-                    Instruction::Jump(4),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(4),
-                    Instruction::StoreVar(3),
-                    Instruction::Jump(-21),
-                    Instruction::Pop(1),
-                    // return result
-                    Instruction::LoadVar(2),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn continue_nested() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: r#"
-                fn Nested() -> int {
-                    while (true) {
-                        while (false) {
-                            continue;
-                        }
-                        if (false) {
-                            continue;
-                        }
-                    }
-                    5
-                }
-            "#,
-            expected: vec![(
-                "Nested",
-                vec![
-                    Instruction::LoadConst(0),
-                    Instruction::JumpIfFalse(15),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(1),
-                    Instruction::JumpIfFalse(4),
-                    Instruction::Pop(1),
-                    Instruction::Jump(1),
-                    Instruction::Jump(-4),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(2),
-                    Instruction::JumpIfFalse(4),
-                    Instruction::Pop(1),
-                    Instruction::Jump(3),
-                    Instruction::Jump(2),
-                    Instruction::Pop(1),
-                    Instruction::Jump(-15),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(3),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn break_nested() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: r#"
-                fn Nested() -> int {
-                    let a = 5;
-                    while (true) {
-                        while (true) {
-                            a = a + 1;
-                            break;
-                        }
-                        a = a + 1;
-                        break;
-                    }
-                    a
-                }
-            "#,
-            expected: vec![(
-                "Nested",
-                vec![
-                    Instruction::LoadConst(0),
-                    Instruction::LoadConst(1),
-                    Instruction::JumpIfFalse(18),
-                    Instruction::Pop(1),
-                    Instruction::LoadConst(2),
-                    Instruction::JumpIfFalse(8),
-                    Instruction::Pop(1),
-                    Instruction::LoadVar(1),
-                    Instruction::LoadConst(3),
-                    Instruction::BinOp(BinOp::Add),
-                    Instruction::StoreVar(1),
-                    Instruction::Jump(3),
-                    Instruction::Jump(-8),
-                    Instruction::Pop(1),
-                    Instruction::LoadVar(1),
-                    Instruction::LoadConst(4),
-                    Instruction::BinOp(BinOp::Add),
-                    Instruction::StoreVar(1),
-                    Instruction::Jump(3),
-                    Instruction::Jump(-18),
-                    Instruction::Pop(1),
-                    Instruction::LoadVar(1),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn builtin_method_call() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: r#"
-                fn main() -> int {
-                    let arr = [1, 2, 3];
-                    arr.len()
-                }
-            "#,
-            expected: vec![(
-                "main",
-                vec![
-                    Instruction::LoadConst(0),
-                    Instruction::LoadConst(1),
-                    Instruction::LoadConst(2),
-                    Instruction::AllocArray(3),
-                    Instruction::LoadGlobal(GlobalIndex::from_raw(3)),
-                    Instruction::LoadVar(1),
-                    // call with one argument (self)
-                    Instruction::Call(1),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn for_loop_sum() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: r#"
-                fn Sum(xs: int[]) -> int {
-                    let result = 0;
-
-                    for (x in xs) {
-                        result += x;
-                    }
-
-                    result
-                }
-                "#,
-            expected: vec![(
-                "Sum",
-                vec![
-                    Instruction::LoadConst(0),
-                    Instruction::LoadVar(1),
-                    Instruction::LoadGlobal(GlobalIndex::from_raw(3)),
-                    Instruction::LoadVar(3),
-                    Instruction::Call(1),
-                    Instruction::LoadConst(0),
-                    Instruction::LoadVar(5),
-                    Instruction::LoadVar(4),
-                    Instruction::CmpOp(CmpOp::Lt),
-                    Instruction::JumpIfFalse(15),
-                    Instruction::Pop(1),
-                    Instruction::LoadVar(3),
-                    Instruction::LoadVar(5),
-                    Instruction::LoadArrayElement,
-                    Instruction::LoadVar(5),
-                    Instruction::LoadConst(1),
-                    Instruction::BinOp(BinOp::Add),
-                    Instruction::StoreVar(5),
-                    Instruction::LoadVar(2),
-                    Instruction::LoadVar(6),
-                    Instruction::BinOp(BinOp::Add),
-                    Instruction::StoreVar(2),
-                    Instruction::Pop(1),
-                    Instruction::Jump(-17),
-                    Instruction::Pop(1),
-                    Instruction::Pop(3),
-                    Instruction::LoadVar(2),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn for_with_break() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: r#"
-                fn ForWithBreak(xs: int[]) -> int {
-                    let result = 0;
-
-                    for (x in xs) {
-                        if (x > 10) {
-                            break;
-                        }
-                        result += x;
-                    }
-
-                    result
-                }
-                "#,
-            expected: vec![(
-                "ForWithBreak",
-                vec![
-                    Instruction::LoadConst(0),
-                    Instruction::LoadVar(1),
-                    Instruction::LoadGlobal(GlobalIndex::from_raw(3)),
-                    Instruction::LoadVar(3),
-                    Instruction::Call(1),
-                    Instruction::LoadConst(0),
-                    Instruction::LoadVar(5),
-                    Instruction::LoadVar(4),
-                    Instruction::CmpOp(CmpOp::Lt),
-                    Instruction::JumpIfFalse(24),
-                    Instruction::Pop(1),
-                    Instruction::LoadVar(3),
-                    Instruction::LoadVar(5),
-                    Instruction::LoadArrayElement,
-                    Instruction::LoadVar(5),
-                    Instruction::LoadConst(1),
-                    Instruction::BinOp(BinOp::Add),
-                    Instruction::StoreVar(5),
-                    Instruction::LoadVar(6),
-                    Instruction::LoadConst(2),
-                    Instruction::CmpOp(CmpOp::Gt),
-                    Instruction::JumpIfFalse(5),
-                    Instruction::Pop(1),
-                    Instruction::Pop(1),
-                    Instruction::Jump(10),
-                    Instruction::Jump(2),
-                    Instruction::Pop(1),
-                    Instruction::LoadVar(2),
-                    Instruction::LoadVar(6),
-                    Instruction::BinOp(BinOp::Add),
-                    Instruction::StoreVar(2),
-                    Instruction::Pop(1),
-                    Instruction::Jump(-26),
-                    Instruction::Pop(1),
-                    Instruction::Pop(3),
-                    Instruction::LoadVar(2),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn for_with_continue() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: r#"
-                fn ForWithContinue(xs: int[]) -> int {
-                    let result = 0;
-
-                    for (x in xs) {
-                        if (x > 10) {
-                            continue;
-                        }
-                        result += x;
-                    }
-
-                    result
-                }
-                "#,
-            expected: vec![(
-                "ForWithContinue",
-                vec![
-                    Instruction::LoadConst(0),
-                    Instruction::LoadVar(1),
-                    Instruction::LoadGlobal(GlobalIndex::from_raw(3)),
-                    Instruction::LoadVar(3),
-                    Instruction::Call(1),
-                    Instruction::LoadConst(0),
-                    Instruction::LoadVar(5),
-                    Instruction::LoadVar(4),
-                    Instruction::CmpOp(CmpOp::Lt),
-                    Instruction::JumpIfFalse(24),
-                    Instruction::Pop(1),
-                    Instruction::LoadVar(3),
-                    Instruction::LoadVar(5),
-                    Instruction::LoadArrayElement,
-                    Instruction::LoadVar(5),
-                    Instruction::LoadConst(1),
-                    Instruction::BinOp(BinOp::Add),
-                    Instruction::StoreVar(5),
-                    Instruction::LoadVar(6),
-                    Instruction::LoadConst(2),
-                    Instruction::CmpOp(CmpOp::Gt),
-                    Instruction::JumpIfFalse(5),
-                    Instruction::Pop(1),
-                    Instruction::Pop(1),
-                    Instruction::Jump(8),
-                    Instruction::Jump(2),
-                    Instruction::Pop(1),
-                    Instruction::LoadVar(2),
-                    Instruction::LoadVar(6),
-                    Instruction::BinOp(BinOp::Add),
-                    Instruction::StoreVar(2),
-                    Instruction::Pop(1),
-                    Instruction::Jump(-26),
-                    Instruction::Pop(1),
-                    Instruction::Pop(3),
-                    Instruction::LoadVar(2),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn for_nested() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: r#"
-                fn NestedFor(as: int[], bs: int[]) -> int {
-
-                    let result = 0;
-
-                    for (a in as) {
-                        for (b in bs) {
-                            result += a * b;
-                        }
-                    }
-
-                    result
-                }
-                "#,
-            expected: vec![(
-                "NestedFor",
-                vec![
-                    Instruction::LoadConst(0),
-                    Instruction::LoadVar(1),
-                    Instruction::LoadGlobal(GlobalIndex::from_raw(3)),
-                    Instruction::LoadVar(4),
-                    Instruction::Call(1),
-                    Instruction::LoadConst(0),
-                    Instruction::LoadVar(6),
-                    Instruction::LoadVar(5),
-                    Instruction::CmpOp(CmpOp::Lt),
-                    Instruction::JumpIfFalse(38),
-                    Instruction::Pop(1),
-                    Instruction::LoadVar(4),
-                    Instruction::LoadVar(6),
-                    Instruction::LoadArrayElement,
-                    Instruction::LoadVar(6),
-                    Instruction::LoadConst(1),
-                    Instruction::BinOp(BinOp::Add),
-                    Instruction::StoreVar(6),
-                    Instruction::LoadVar(2),
-                    Instruction::LoadGlobal(GlobalIndex::from_raw(3)),
-                    Instruction::LoadVar(8),
-                    Instruction::Call(1),
-                    Instruction::LoadConst(0),
-                    Instruction::LoadVar(10),
-                    Instruction::LoadVar(9),
-                    Instruction::CmpOp(CmpOp::Lt),
-                    Instruction::JumpIfFalse(17),
-                    Instruction::Pop(1),
-                    Instruction::LoadVar(8),
-                    Instruction::LoadVar(10),
-                    Instruction::LoadArrayElement,
-                    Instruction::LoadVar(10),
-                    Instruction::LoadConst(1),
-                    Instruction::BinOp(BinOp::Add),
-                    Instruction::StoreVar(10),
-                    Instruction::LoadVar(3),
-                    Instruction::LoadVar(7),
-                    Instruction::LoadVar(11),
-                    Instruction::BinOp(BinOp::Mul),
-                    Instruction::BinOp(BinOp::Add),
-                    Instruction::StoreVar(3),
-                    Instruction::Pop(1),
-                    Instruction::Jump(-19),
-                    Instruction::Pop(1),
-                    Instruction::Pop(3),
-                    Instruction::Pop(1),
-                    Instruction::Jump(-40),
-                    Instruction::Pop(1),
-                    Instruction::Pop(3),
-                    Instruction::LoadVar(3),
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    mod return_stmt {
-        use super::*;
-
-        #[test]
-        fn early_return() -> anyhow::Result<()> {
-            assert_compiles(Program {
-                source: "
-                fn EarlyReturn(x: int) -> int {
-                  if (x == 42) { return 1; }
-
-                  x + 5
-                }
-            ",
-                expected: vec![(
-                    "EarlyReturn",
-                    vec![
-                        Instruction::LoadVar(1),   // x
-                        Instruction::LoadConst(0), // 42
-                        Instruction::CmpOp(CmpOp::Eq),
-                        Instruction::JumpIfFalse(5), // to 8
-                        Instruction::Pop(1),
-                        Instruction::LoadConst(1), // 1
-                        Instruction::Return,
-                        Instruction::Jump(2), // to 9
-                        Instruction::Pop(1),
-                        Instruction::LoadVar(1),   // x
-                        Instruction::LoadConst(2), // 5
-                        Instruction::BinOp(BinOp::Add),
-                        Instruction::Return,
-                    ],
-                )],
-            })
-        }
-
-        #[test]
-        fn with_stack() -> anyhow::Result<()> {
-            assert_compiles(Program {
-                source: "
-                fn WithStack(x: int) -> int {
-                  let a = 1;
-
-                  // NOTE: currently there's no empty returns.
-
-                  if (a == 0) { return 0; }
-
-                  {
-                     let b = 1;
-                     if (a != b) {
-                        return 0;
-                     }
-                  }
-
-                  {
-                     let c = 2;
-                     let b = 3;
-                     while (b != c) {
-                        if (true) {
-                           return 0;
-                        }
-                     }
-                  }
-
-                   7
-                }
-            ",
-                expected: vec![(
-                    "WithStack",
-                    vec![
-                        Instruction::LoadConst(0), // 1
-                        Instruction::LoadVar(2),   // a
-                        Instruction::LoadConst(1), // 0
-                        Instruction::CmpOp(CmpOp::Eq),
-                        Instruction::JumpIfFalse(5), // to 9
-                        Instruction::Pop(1),
-                        Instruction::LoadConst(2), // 0
-                        Instruction::Return,
-                        Instruction::Jump(2), // to 10
-                        Instruction::Pop(1),
-                        Instruction::LoadConst(3), // 1
-                        Instruction::LoadVar(2),   // a
-                        Instruction::LoadVar(3),   // b
-                        Instruction::CmpOp(CmpOp::NotEq),
-                        Instruction::JumpIfFalse(5), // to 19
-                        Instruction::Pop(1),
-                        Instruction::LoadConst(4), // 0
-                        Instruction::Return,
-                        Instruction::Jump(2), // to 20
-                        Instruction::Pop(1),
-                        Instruction::Pop(1),
-                        Instruction::LoadConst(5), // 2
-                        Instruction::LoadConst(6), // 3
-                        Instruction::LoadVar(4),   // b
-                        Instruction::LoadVar(3),   // c
-                        Instruction::CmpOp(CmpOp::NotEq),
-                        Instruction::JumpIfFalse(10), // to 36
-                        Instruction::Pop(1),
-                        Instruction::LoadConst(7),   // true
-                        Instruction::JumpIfFalse(5), // to 34
-                        Instruction::Pop(1),
-                        Instruction::LoadConst(8), // 0
-                        Instruction::Return,
-                        Instruction::Jump(2), // to 35
-                        Instruction::Pop(1),
-                        Instruction::Jump(-12), // to 23
-                        Instruction::Pop(1),
-                        Instruction::Pop(2),
-                        Instruction::LoadConst(9), // 7
-                        Instruction::Return,
-                    ],
-                )],
-            })
-        }
-    }
-
-    #[test]
-    fn assert_statement_ok() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: "
-                fn assertOk() -> int {
-                    assert 2 + 2 == 4;
-                    3
-                }
-            ",
-            expected: vec![(
-                "assertOk",
-                vec![
-                    Instruction::LoadConst(0), // 2
-                    Instruction::LoadConst(1), // 2
-                    Instruction::BinOp(BinOp::Add),
-                    Instruction::LoadConst(2), // 4
-                    Instruction::CmpOp(CmpOp::Eq),
-                    Instruction::Assert,
-                    Instruction::LoadConst(3), // 3
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn assert_statement_not_ok() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: "
-                fn assertNotOk() -> int {
-                    assert 3 == 1;
-                    2
-                }
-            ",
-            expected: vec![(
-                "assertNotOk",
-                vec![
-                    Instruction::LoadConst(0), // 3
-                    Instruction::LoadConst(1), // 1
-                    Instruction::CmpOp(CmpOp::Eq),
-                    Instruction::Assert,
-                    Instruction::LoadConst(2), // 2
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn field_assignment_compound_add_bytecode() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: "
-                class Counter {
-                    value int
-                }
-
-                function incrementCounter(c: Counter) -> int {
-                    c.value += 10;
-                    c.value
-                }
-            ",
-            expected: vec![(
-                "incrementCounter",
-                vec![
-                    // c.value += 10
-                    Instruction::LoadVar(1),        // Load c
-                    Instruction::Copy(0),           // Duplicate c reference
-                    Instruction::LoadField(0),      // Load c.value
-                    Instruction::LoadConst(0),      // Load 10
-                    Instruction::BinOp(BinOp::Add), // Add
-                    Instruction::StoreField(0),     // Store back to c.value
-                    // c.value
-                    Instruction::LoadVar(1),   // Load c
-                    Instruction::LoadField(0), // Load c.value
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn nested_field_read_bytecode() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: "
-                class Inner {
-                    value int
-                }
-                class Outer {
-                    inner Inner
-                }
-
-                function main() -> int {
-                    let o = Outer { inner: Inner { value: 42 } };
-                    o.inner.value
-                }
-            ",
-            expected: vec![(
-                "main",
-                vec![
-                    // Create Outer { inner: Inner { value: 42 } }
-                    Instruction::AllocInstance(ObjectIndex::from_raw(3)), // Outer class
-                    Instruction::Copy(0),                                 // Copy Outer instance
-                    // Create Inner inline
-                    Instruction::AllocInstance(ObjectIndex::from_raw(2)), // Inner class
-                    Instruction::Copy(0),                                 // Copy Inner instance
-                    Instruction::LoadConst(0),                            // 42
-                    Instruction::StoreField(0),                           // Inner.value = 42
-                    Instruction::StoreField(0), // Outer.inner = Inner instance
-                    // o.inner.value
-                    Instruction::LoadVar(1),   // Load o
-                    Instruction::LoadField(0), // Load o.inner (returns Inner)
-                    Instruction::LoadField(0), // Load inner.value (returns 42)
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn nested_object_construction_bytecode() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: "
-                class Inner {
-                    x int
-                    y int
-                }
-                class Outer {
-                    inner Inner
-                    value int
-                }
-
-                function main() -> int {
-                    let o = Outer {
-                        inner: Inner { x: 10, y: 20 },
-                        value: 30
-                    };
-                    o.value
-                }
-            ",
-            expected: vec![(
-                "main",
-                vec![
-                    // Outer constructor
-                    Instruction::AllocInstance(ObjectIndex::from_raw(3)), // Outer
-                    Instruction::Copy(0),                                 // Copy Outer instance
-                    // Nested Inner construction
-                    Instruction::AllocInstance(ObjectIndex::from_raw(2)), // Inner
-                    Instruction::Copy(0),                                 // Copy Inner instance
-                    Instruction::LoadConst(0),                            // 10
-                    Instruction::StoreField(0),                           // x = 10
-                    Instruction::Copy(0),       // Copy Inner instance again
-                    Instruction::LoadConst(1),  // 20
-                    Instruction::StoreField(1), // y = 20
-                    Instruction::StoreField(0), // Outer.inner = Inner
-                    Instruction::Copy(0),       // Copy Outer instance
-                    Instruction::LoadConst(2),  // 30
-                    Instruction::StoreField(1), // Outer.value = 30
-                    // o.value
-                    Instruction::LoadVar(1),   // o
-                    Instruction::LoadField(1), // value
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn nested_field_assignment_bytecode() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: "
-                class Inner {
-                    value int
-                }
-                class Outer {
-                    inner Inner
-                }
-
-                function setNestedValue(i: Inner, o: Outer) -> int {
-                    o.inner.value = 99;
-                    o.inner.value
-                }
-            ",
-            expected: vec![(
-                "setNestedValue",
-                vec![
-                    // o.inner.value = 99
-                    Instruction::LoadVar(2),    // Load o
-                    Instruction::LoadField(0),  // Load o.inner (returns Inner object)
-                    Instruction::LoadConst(0),  // Load 99
-                    Instruction::StoreField(0), // Store to inner.value
-                    // o.inner.value
-                    Instruction::LoadVar(2),   // Load o
-                    Instruction::LoadField(0), // Load o.inner
-                    Instruction::LoadField(0), // Load inner.value
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn nested_field_assignment_compound_bytecode() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: "
-                class Inner {
-                    value int
-                }
-                class Outer {
-                    inner Inner
-                }
-
-                function incrementNestedValue(o: Outer) -> int {
-                    o.inner.value += 10;
-                    o.inner.value
-                }
-            ",
-            expected: vec![(
-                "incrementNestedValue",
-                vec![
-                    // o.inner.value += 10
-                    Instruction::LoadVar(1),        // Load o
-                    Instruction::LoadField(0),      // Load o.inner (returns Inner object)
-                    Instruction::Copy(0),           // Duplicate inner reference
-                    Instruction::LoadField(0),      // Load inner.value
-                    Instruction::LoadConst(0),      // Load 10
-                    Instruction::BinOp(BinOp::Add), // Add
-                    Instruction::StoreField(0),     // Store back to inner.value
-                    // o.inner.value
-                    Instruction::LoadVar(1),   // Load o
-                    Instruction::LoadField(0), // Load o.inner
-                    Instruction::LoadField(0), // Load inner.value
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    #[test]
-    fn field_assignment_simple_bytecode() -> anyhow::Result<()> {
-        assert_compiles(Program {
-            source: "
-                class Data {
-                    value int
-                }
-
-                function setDataValue(d: Data) -> int {
-                    d.value = 42;
-                    d.value
-                }
-            ",
-            expected: vec![(
-                "setDataValue",
-                vec![
-                    // d.value = 42
-                    Instruction::LoadVar(1),    // Load d
-                    Instruction::LoadConst(0),  // Load 42
-                    Instruction::StoreField(0), // Store to d.value
-                    // d.value
-                    Instruction::LoadVar(1),   // Load d
-                    Instruction::LoadField(0), // Load d.value
-                    Instruction::Return,
-                ],
-            )],
-        })
-    }
-
-    mod maps {
-        use super::*;
-
-        #[test]
-        fn create_and_access() -> anyhow::Result<()> {
-            assert_compiles(Program {
-                source: r#"fn CreateMap() -> map<string, string> {
-    { hello "world" }
-}
-fn UseMap() -> string {
-    let map = CreateMap();
-    map["hello"]
-}"#,
-                expected: vec![
-                    (
-                        "CreateMap",
-                        vec![
-                            Instruction::LoadConst(0),
-                            Instruction::LoadConst(1),
-                            Instruction::AllocMap(1),
-                            Instruction::Return,
-                        ],
-                    ),
-                    (
-                        "UseMap",
-                        vec![
-                            Instruction::LoadGlobal(GlobalIndex::from_raw(0)),
-                            Instruction::Call(0),
-                            Instruction::LoadVar(1),
-                            Instruction::LoadConst(0),
-                            Instruction::LoadMapElement,
-                            Instruction::Return,
-                        ],
-                    ),
-                ],
-            })
-        }
-
-        #[test]
-        fn access_no_key() -> anyhow::Result<()> {
-            assert_compiles(Program {
-                source: r#"
-fn CreateMap() -> map<string, string> {
-    { hello "world" }
-}
-
-fn UseMapNoKey() -> string {
-    let map = CreateMap();
-    map["world"]
-}"#,
-                expected: vec![
-                    (
-                        "CreateMap",
-                        vec![
-                            Instruction::LoadConst(0),
-                            Instruction::LoadConst(1),
-                            Instruction::AllocMap(1),
-                            Instruction::Return,
-                        ],
-                    ),
-                    (
-                        "UseMapNoKey",
-                        vec![
-                            Instruction::LoadGlobal(GlobalIndex::from_raw(0)),
-                            Instruction::Call(0),
-                            Instruction::LoadVar(1),
-                            Instruction::LoadConst(0),
-                            Instruction::LoadMapElement,
-                            Instruction::Return,
-                        ],
-                    ),
-                ],
-            })
-        }
-
-        #[test]
-        fn contains() -> anyhow::Result<()> {
-            assert_compiles(Program {
-                source: r#"
-                    fn CreateMapJSON() -> map<string, string> {
-                        {"hello": "world"}
-                    }
-                    fn UseMapContains() -> string {
-                        let map = CreateMapJSON();
-                        if (map.contains("hello")) {
-                            map["hello"]
-                        } else {
-                            "hi"
-                        }
-                    }
-                "#,
-                expected: vec![
-                    (
-                        "CreateMapJSON",
-                        vec![
-                            Instruction::LoadConst(0),
-                            Instruction::LoadConst(1),
-                            Instruction::AllocMap(1),
-                            Instruction::Return,
-                        ],
-                    ),
-                    (
-                        "UseMapContains",
-                        vec![
-                            Instruction::LoadGlobal(GlobalIndex::from_raw(0)),
-                            Instruction::Call(0),
-                            Instruction::LoadGlobal(GlobalIndex::from_raw(6)),
-                            Instruction::LoadVar(1),
-                            Instruction::LoadConst(0),
-                            Instruction::Call(2),
-                            Instruction::JumpIfFalse(6),
-                            Instruction::Pop(1),
-                            Instruction::LoadVar(1),
-                            Instruction::LoadConst(1),
-                            Instruction::LoadMapElement,
-                            Instruction::Jump(3),
-                            Instruction::Pop(1),
-                            Instruction::LoadConst(2),
-                            Instruction::Return,
-                        ],
-                    ),
-                ],
-            })
-        }
-
-        #[test]
-        fn modify() -> anyhow::Result<()> {
-            assert_compiles(Program {
-                source: r#"
-                    fn EditMapKey() -> int {
-                        let map = { hi 123 };
-
-                        map["hi"] = 42 - 4;
-                        map["hi"] += 4;
-
-                        map["hi"]
-
-                    }
-                "#,
-                expected: vec![(
-                    "EditMapKey",
-                    vec![
-                        Instruction::LoadConst(0),
-                        Instruction::LoadConst(1),
-                        Instruction::AllocMap(1),
-                        Instruction::LoadVar(1),
-                        Instruction::LoadConst(2),
-                        Instruction::LoadConst(3),
-                        Instruction::LoadConst(4),
-                        Instruction::BinOp(BinOp::Sub),
-                        Instruction::StoreMapElement,
-                        Instruction::LoadVar(1),
-                        Instruction::LoadConst(5),
-                        Instruction::Copy(1),
-                        Instruction::Copy(1),
-                        Instruction::LoadMapElement,
-                        Instruction::LoadConst(6),
-                        Instruction::BinOp(BinOp::Add),
-                        Instruction::StoreMapElement,
-                        Instruction::LoadVar(1),
-                        Instruction::LoadConst(7),
-                        Instruction::LoadMapElement,
-                        Instruction::Return,
-                    ],
-                )],
-            })
-        }
-
-        #[test]
-        fn len() -> anyhow::Result<()> {
-            assert_compiles(Program {
-                source: r#"
-                    fn Len() -> int {
-                        let map = {
-                            hi 123
-                            it_works 456
-                        };
-                        map.len()
-                    }
-                "#,
-                expected: vec![(
-                    "Len",
-                    vec![
-                        Instruction::LoadConst(0),
-                        Instruction::LoadConst(1),
-                        Instruction::LoadConst(2),
-                        Instruction::LoadConst(3),
-                        Instruction::AllocMap(2),
-                        Instruction::LoadGlobal(GlobalIndex::from_raw(4)),
-                        Instruction::LoadVar(1),
-                        Instruction::Call(1),
-                        Instruction::Return,
-                    ],
-                )],
-            })
-        }
-    }
-
-    mod enums {
-        use super::*;
-
-        #[test]
-        fn return_enum_variant() -> anyhow::Result<()> {
-            assert_compiles(Program {
-                source: r#"
-                    enum Shape {
-                        Square
-                        Rectangle
-                        Circle
-                    }
-
-                    fn main() -> Shape {
-                        Shape.Rectangle
-                    }
-                "#,
-                expected: vec![(
-                    "main",
-                    vec![
-                        Instruction::LoadConst(0),
-                        Instruction::AllocVariant(ObjectIndex::from_raw(3)),
-                        Instruction::Return,
-                    ],
-                )],
-            })
-        }
-
-        #[test]
-        fn assign_enum_variant() -> anyhow::Result<()> {
-            assert_compiles(Program {
-                source: r#"
-                    enum Shape {
-                        Square
-                        Rectangle
-                        Circle
-                    }
-
-                    fn main() -> Shape {
-                        let s = Shape.Rectangle;
-                        s
-                    }
-                "#,
-                expected: vec![(
-                    "main",
-                    vec![
-                        Instruction::LoadConst(0),
-                        Instruction::AllocVariant(ObjectIndex::from_raw(3)),
-                        Instruction::LoadVar(1),
-                        Instruction::Return,
-                    ],
-                )],
-            })
-        }
-    }
-}
